Clean up a few things, and experiment with code coverage

+++ src/main/java/net/daboross/bukkitdev/skywars/api/arenaconfig/SkyArenaConfig.java
@@ -133,7 +133,7 @@ public class SkyArenaConfig extends Parentable<SkyArenaConfig> implements SkyAre
                 return parent.getNumTeams();
             }
         } else {
-            return rawNumTeams.intValue();
+            return rawNumTeams;
         }
     }
